requirements.txt
beautifulsoup4==4.12.2
imgkit==1.2.3
mail-parser==3.15.0
openpyxl==3.1.2
python-dotenv==1.0.0
img2pdf==0.4.4